converted function to class
