+++ requirements.txt
@@ -1,5 +1,5 @@
 Faker==23.2.1
 pyspark==3.5.0
-wheel==0.42.0
-twine==5.0.0
-setuptools==69.1.1
+#wheel==0.42.0
+#twine==5.0.0
+#setuptools==69.1.1
