+++ tests/test_stats.py
@@ -1,6 +1,8 @@
-from pyspark_utils.stats import Stats
+from pyspark_utility.object_size_calculator import ObjectSizeCalculator
 from pyspark.sql import DataFrame, SparkSession
 
+
+#
 # spark = (
 #   SparkSession
 #     .builder
@@ -8,8 +10,8 @@ from pyspark.sql import DataFrame, SparkSession
 #     .master("local[*]")
 #     .getOrCreate()
 # )
-
-
+#
+#
 # obj = [1,2,3,4]
-# size_in_bytes = Stats.get_size_for_machine(obj,spark=spark)
+# size_in_bytes = ObjectSizeCalculator.get_size_for_machine(obj,spark=spark)
 # assert size_in_bytes==144, True
